tf_job/requirement_cpu.txt
flask
tensorflow
pandas
docker
flask_pymongo
mlflow
bert_tensorflow
sklearn
tensorflow_hub
